+++ tomcat/src/main/java/org/apache/coyote/http11/controller/AbstractController.java
@@ -4,7 +4,7 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 
-import org.apache.coyote.http11.HttpRequest;
+import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.HttpResponse;
 
 public abstract class AbstractController implements Controller {
