+++ tomcat/src/main/java/org/apache/coyote/http11/controller/Controller.java
@@ -1,6 +1,6 @@
 package org.apache.coyote.http11.controller;
 
-import org.apache.coyote.http11.HttpRequest;
+import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.HttpResponse;
 
 public interface Controller {
