+++ tomcat/src/main/java/org/apache/coyote/http11/controller/FaviconController.java
@@ -2,7 +2,7 @@ package org.apache.coyote.http11.controller;
 
 import java.io.IOException;
 
-import org.apache.coyote.http11.HttpRequest;
+import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.HttpResponse;
 
 public class FaviconController extends AbstractController {
