+++ tomcat/src/main/java/org/apache/coyote/http11/controller/RegisterController.java
@@ -3,7 +3,7 @@ package org.apache.coyote.http11.controller;
 import java.io.IOException;
 import java.util.Map;
 
-import org.apache.coyote.http11.HttpRequest;
+import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.HttpResponse;
 
 import com.techcourse.db.InMemoryUserRepository;
