+++ tomcat/src/test/java/org/apache/coyote/http11/controller/LoginControllerTest.java
@@ -12,7 +12,7 @@ import java.nio.file.Files;
 
 import org.apache.catalina.Session;
 import org.apache.catalina.SessionManager;
-import org.apache.coyote.http11.HttpRequest;
+import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.HttpResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
