+++ tomcat/src/test/java/org/apache/coyote/http11/controller/StaticResourceControllerTest.java
@@ -9,7 +9,7 @@ import java.io.OutputStream;
 import java.net.URL;
 import java.nio.file.Files;
 
-import org.apache.coyote.http11.HttpRequest;
+import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.HttpResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
